+++ tests/unit/hooks.test.ts
@@ -290,6 +290,22 @@ describe('Hook registry', () => {
 		expect(listener).toBeCalledWith(visit, undefined, undefined);
 	});
 
+	it('should allow replacing original handler only once', async () => {
+		const swup = new Swup();
+		const customHandler = vi.fn();
+		const defaultHandler = vi.fn();
+
+		swup.hooks.on('enable', customHandler, { replace: true, once: true });
+
+		await swup.hooks.call('enable', undefined, undefined, defaultHandler);
+		expect(defaultHandler).toBeCalledTimes(0);
+		expect(customHandler).toBeCalledTimes(1);
+
+		await swup.hooks.call('enable', undefined, undefined, defaultHandler);
+		expect(defaultHandler).toBeCalledTimes(1);
+		expect(customHandler).toBeCalledTimes(1);
+	});
+
 	it('should trigger hook handler with visit and args', async () => {
 		const swup = new SwupWithPublicVisitMethods();
 		const handler: HookHandler<'history:popstate'> = vi.fn();
@@ -377,14 +393,16 @@ describe('Hook registry', () => {
 			hooks: {
 				'visit:start': handler,
 				'visit:start.before': handler,
-				'visit:start.once': handler
+				'visit:start.once': handler,
+				'visit:start.once.before': handler
 			}
 		});
 
-		expect(hookSpy).toBeCalledTimes(3);
+		expect(hookSpy).toBeCalledTimes(4);
 		expect(hookSpy).toBeCalledWith('visit:start', handler, {});
 		expect(hookSpy).toBeCalledWith('visit:start', handler, { before: true });
 		expect(hookSpy).toBeCalledWith('visit:start', handler, { once: true });
+		expect(hookSpy).toBeCalledWith('visit:start', handler, { once: true, before: true });
 	});
 });